90% win rate-Sarah

+++ player.hpp
@@ -14,7 +14,7 @@ public:
     ~Player();
 
     Move *doMove(Move *opponentsMove, int msLeft);
-    int calcScore(Board * b, Move *m, Side side);
+    double calcScore(Board * b, Move *m, Side side);
     // Flag to tell if the player is running within the test_minimax context
     bool testingMinimax;
     Side ourSide;
